refactor: make the popup logic more general
