+++ internal/popup/popup.go
@@ -3,45 +3,30 @@ package popup
 import (
 	"strings"
 
-	"github.com/TypicalAM/goread/internal/colorscheme"
-	"github.com/TypicalAM/goread/internal/rss"
-	"github.com/charmbracelet/bubbles/textinput"
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/charmbracelet/lipgloss"
 	"github.com/muesli/ansi"
 )
 
-// ChosenCategoryMsg is the message displayed when a category is successfully chosen.
-type ChosenCategoryMsg struct {
-	Name string
+// Popup is the popup window interface. In can be implemented in other packages and use the `Default` popup to overlay the content
+// on top of the background.
+type Popup interface {
+	tea.Model
 }
 
-// focusedField is the field that is currently focused.
-type focusedField int
-
-const (
-	allField focusedField = iota
-	downloadedField
-	newCategoryField
-)
-
-// Popup is a popup window allowing the user to select an item from a list of items.
-type Popup struct {
-	ogSection   []string
-	section     []string
-	width       int
-	height      int
-	prefix      string
-	suffix      string
-	colors      colorscheme.Colorscheme
-	textInput   textinput.Model
-	focused     focusedField
-	startCol    int
-	renderStyle lipgloss.Style
+// Default is a default popup window.
+type Default struct {
+	ogSection []string
+	section   []string
+	width     int
+	height    int
+	prefix    string
+	suffix    string
+	startCol  int
 }
 
-// New creates a new popup window.
-func New(colors colorscheme.Colorscheme, bgRaw string, width, height int) Popup {
+// New creates a new default popup window.
+func New(bgRaw string, width, height int) Default {
 	bg := strings.Split(bgRaw, "\n")
 	bgWidth := ansi.PrintableRuneWidth(bg[0])
 	bgHeight := len(bg)
@@ -56,9 +41,7 @@ func New(colors colorscheme.Colorscheme, bgRaw string, width, height int) Popup 
 	suffix := strings.Join(bg[startRow+height:], "\n")
 	copy(ogSection, bg[startRow:startRow+height])
 
-	text := textinput.New()
-
-	return Popup{
+	return Default{
 		ogSection: ogSection,
 		section:   section,
 		width:     width,
@@ -66,133 +49,19 @@ func New(colors colorscheme.Colorscheme, bgRaw string, width, height int) Popup 
 		prefix:    prefix,
 		suffix:    suffix,
 		startCol:  startCol,
-		colors:    colors,
-		textInput: text,
-		renderStyle: lipgloss.NewStyle().
-			Foreground(lipgloss.Color("#FFFFFF")).
-			Width(width - 2).
-			Height(height - 2).
-			Border(lipgloss.NormalBorder()).
-			BorderForeground(colors.Color1),
 	}
 }
 
-// Init the popup window.
-func (p Popup) Init() tea.Cmd {
-	return textinput.Blink
-}
-
-// Update the popup window.
-func (p Popup) Update(msg tea.Msg) (Popup, tea.Cmd) {
-	var cmds []tea.Cmd
-
-	if msg, ok := msg.(tea.KeyMsg); ok {
-		switch msg.String() {
-		case "down", "j":
-			switch p.focused {
-			case allField:
-				p.focused = downloadedField
-			case downloadedField:
-				p.focused = newCategoryField
-				cmds = append(cmds, p.textInput.Focus())
-			case newCategoryField:
-				p.focused = allField
-				p.textInput.Blur()
-			}
-
-		case "up", "k":
-			switch p.focused {
-			case allField:
-				p.focused = newCategoryField
-				cmds = append(cmds, p.textInput.Focus())
-			case downloadedField:
-				p.focused = allField
-			case newCategoryField:
-				p.focused = downloadedField
-				p.textInput.Blur()
-			}
-
-		case "enter":
-			switch p.focused {
-			case allField:
-				return p, confirmCategory(rss.AllFeedsName)
-
-			case downloadedField:
-				return p, confirmCategory(rss.DownloadedFeedsName)
-
-			case newCategoryField:
-				// TODO: Validate the name
-				return p, confirmCategory(p.textInput.Value())
-			}
-		}
-	}
-
-	if p.textInput.Focused() {
-		var cmd tea.Cmd
-		p.textInput, cmd = p.textInput.Update(msg)
-		cmds = append(cmds, cmd)
-	}
-
-	return p, tea.Batch(cmds...)
-}
-
-// Render the popup window.
-func (p Popup) View() string {
-	// Question
-	headingStyle := lipgloss.NewStyle().
-		Margin(2, 2).
-		Width(p.width - 2).
-		Align(lipgloss.Center).
-		Italic(true)
-
-	question := headingStyle.Render("Which category do you want to add?")
-
-	choices := []string{"All", "Downloaded", "New Category"}
-	descriptions := []string{
-		"All the feeds from all the categories",
-		"Downloaded articles",
-		"",
-	}
-	choiceSectionStyle := lipgloss.NewStyle().
-		Padding(2).
-		Width(p.width - 2).
-		Height(10)
-
-	choiceStyle := lipgloss.NewStyle().
-		PaddingLeft(2).
-		MarginBottom(1).
-		Border(lipgloss.RoundedBorder(), false, false, false, true)
-
-	selectedChoiceStyle := choiceStyle.Copy().
-		BorderForeground(p.colors.Color4)
-
-	renderedChoices := make([]string, len(choices))
-	if p.focused == allField {
-		renderedChoices[0] = selectedChoiceStyle.Render(lipgloss.JoinVertical(lipgloss.Top, choices[0], descriptions[0]))
-	} else {
-		renderedChoices[0] = choiceStyle.Render(lipgloss.JoinVertical(lipgloss.Top, choices[0], descriptions[0]))
-	}
-
-	if p.focused == downloadedField {
-		renderedChoices[1] = selectedChoiceStyle.Render(lipgloss.JoinVertical(lipgloss.Top, choices[1], descriptions[1]))
-	} else {
-		renderedChoices[1] = choiceStyle.Render(lipgloss.JoinVertical(lipgloss.Top, choices[1], descriptions[1]))
-	}
-
-	if p.focused == newCategoryField {
-		renderedChoices[2] = selectedChoiceStyle.Render(lipgloss.JoinVertical(lipgloss.Top, choices[2], p.textInput.View()))
-	} else {
-		renderedChoices[2] = choiceStyle.Render(lipgloss.JoinVertical(lipgloss.Top, choices[2], p.textInput.View()))
-	}
-
-	toBox := choiceSectionStyle.Render(lipgloss.JoinVertical(lipgloss.Top, renderedChoices...))
-	popup := lipgloss.JoinVertical(lipgloss.Top, question, toBox)
-	popupSplit := strings.Split(p.renderStyle.Render(popup), "\n")
+// Overlay overlays the given text on top of the background.
+func (p Default) Overlay(text string) string {
+	// TODO: Add a padding guardrail
+	lines := strings.Split(text, "\n")
 
 	// Overlay the background with the styled text.
+	// TODO: Use a string builder
 	for i, text := range p.ogSection {
 		p.section[i] = text[:findPrintableIndex(text, p.startCol)] +
-			popupSplit[i] +
+			lines[i] +
 			text[findPrintableIndex(text, p.startCol+p.width):]
 	}
 
@@ -204,6 +73,16 @@ func (p Popup) View() string {
 	)
 }
 
+// Width returns the width of the popup window.
+func (p Default) Width() int {
+	return p.width
+}
+
+// Height returns the height of the popup window.
+func (p Default) Height() int {
+	return p.height
+}
+
 // findPrintableIndex finds the index of the last printable rune at the given index.
 func findPrintableIndex(str string, index int) int {
 	for i := len(str) - 1; i >= 0; i-- {
@@ -211,12 +90,6 @@ func findPrintableIndex(str string, index int) int {
 			return i
 		}
 	}
-	return -1
-}
 
-// confirmCategory returns a tea.Cmd which relays the message to the browser.
-func confirmCategory(name string) tea.Cmd {
-	return func() tea.Msg {
-		return ChosenCategoryMsg{Name: name}
-	}
+	return -1
 }
